Refactor JSON-LD emitters > moved duplicate code to parent class
