+++ shared/src/main/scala/amf/core/internal/plugins/document/graph/emitter/FlattenedJsonLdEmitter.scala
@@ -2,28 +2,20 @@ package amf.core.internal.plugins.document.graph.emitter
 
 import amf.core.client.scala.config
 import amf.core.client.scala.config.RenderOptions
-import amf.core.internal.annotations._
-import amf.core.internal.metamodel.Type.{Any, Array, ArrayLike, Bool, EncodedIri, Iri, LiteralUri, SortedArray, Str}
+import amf.core.client.scala.model.document.{BaseUnit, EncodesModel, SourceMap}
+import amf.core.client.scala.model.domain._
+import amf.core.client.scala.model.domain.extensions.DomainExtension
+import amf.core.client.scala.vocabulary.{Namespace, NamespaceAliases}
+import amf.core.internal.metamodel.Type._
 import amf.core.internal.metamodel._
 import amf.core.internal.metamodel.document.{BaseUnitModel, FragmentModel, ModuleModel, SourceMapModel}
 import amf.core.internal.metamodel.domain.extensions.DomainExtensionModel
-import amf.core.internal.metamodel.domain.{DomainElementModel, LinkableElementModel, ShapeModel}
-import amf.core.client.scala.model.DataType
-import amf.core.client.scala.model.document.{BaseUnit, EncodesModel, SourceMap}
-import amf.core.client.scala.model.domain.DataNodeOps.adoptTree
-import amf.core.client.scala.model.domain._
-import amf.core.client.scala.model.domain.extensions.DomainExtension
 import amf.core.internal.parser.domain.{Annotations, FieldEntry, Value}
-import amf.core.client.scala.vocabulary.{Namespace, NamespaceAliases, ValueType}
 import amf.core.internal.plugins.document.graph.JsonLdKeywords
 import amf.core.internal.plugins.document.graph.emitter.flattened.utils.{Emission, EmissionQueue, Metadata}
-import amf.core.internal.plugins.document.graph.emitter.utils.SourceMapsAllowList
-import org.mulesoft.common.time.SimpleDateTime
 import org.yaml.builder.DocBuilder
 import org.yaml.builder.DocBuilder.{Entry, Part, SType, Scalar}
 
-import scala.collection.mutable
-import scala.collection.mutable.ListBuffer
 import scala.language.implicitConversions
 
 object FlattenedJsonLdEmitter {
@@ -35,7 +27,7 @@ object FlattenedJsonLdEmitter {
       namespaceAliases: NamespaceAliases = Namespace.defaultAliases,
       fieldProvision: ApplicableMetaFieldRenderProvider
   ): Boolean = {
-    implicit val ctx: GraphEmitterContext =
+    implicit val ctx: FlattenedGraphEmitterContext =
       FlattenedGraphEmitterContext(unit, renderOptions, namespaceAliases = namespaceAliases)
     new FlattenedJsonLdEmitter[T](builder, renderOptions, fieldProvision).root(unit)
     true
@@ -46,13 +38,12 @@ class FlattenedJsonLdEmitter[T](
     val builder: DocBuilder[T],
     val options: RenderOptions,
     val fieldProvision: ApplicableMetaFieldRenderProvider
-)(implicit ctx: GraphEmitterContext)
-    extends CommonEmitter
+)(implicit ctx: FlattenedGraphEmitterContext)
+    extends CommonEmitter[T, FlattenedGraphEmitterContext](options)
     with MetaModelTypeMapping {
 
   val pending: EmissionQueue[T] = EmissionQueue()
   var root: Part[T]             = _
-  val allowList: List[String]   = SourceMapsAllowList()
 
   def root(unit: BaseUnit): Unit = {
     builder.obj { ob =>
@@ -131,34 +122,6 @@ class FlattenedJsonLdEmitter[T](
       case e: EncodesModel => Option(e.encodes).exists(_.id == e.id)
       case _               => false
     }
-  }
-
-  private def emitDeclarations(b: Entry[T], id: String, sources: SourceMap): Unit = {
-    if (ctx.declared.nonEmpty) {
-      val v   = Value(AmfArray(ctx.declared), Annotations())
-      val f   = ModuleModel.Declares
-      val url = ctx.emitIri(f.value.iri())
-      ctx.emittingDeclarations(true)
-      b.entry(
-          url,
-          value(f.`type`, v, id, sources.property(url), _)
-      )
-    }
-    ctx.emittingDeclarations(false)
-  }
-
-  private def emitReferences(b: Entry[T], id: String, sources: SourceMap): Unit = {
-    if (ctx.referenced.nonEmpty) {
-      val v   = Value(AmfArray(ctx.referenced), Annotations())
-      val f   = ModuleModel.References
-      val url = ctx.emitIri(f.value.iri())
-      ctx.emittingReferences(true)
-      b.entry(
-          url,
-          value(f.`type`, v, id, sources.property(url), _)
-      )
-    }
-    ctx.emittingReferences(false)
   }
 
   def queueObjectFieldValues(amfObject: AmfObject, filter: (Field, Value) => Boolean = (_, _) => true): Unit = {
@@ -300,69 +263,10 @@ class FlattenedJsonLdEmitter[T](
     modelFields.foreach { f =>
       emitStaticField(f, element, id, sources, b)
     }
+
   }
 
-  private def emitStaticField(field: Field, element: AmfObject, id: String, sources: SourceMap, b: Entry[T]): Unit = {
-    element.fields.entryJsonld(field) match {
-      case Some(FieldEntry(f, v)) =>
-        val url = ctx.emitIri(f.value.iri())
-        b.entry(
-            url,
-            value(f.`type`, v, id, sources.property(url), _)
-        )
-      case None => // Missing field
-    }
-  }
-
-  private def createCustomExtensions(element: AmfObject, b: Entry[T]): Unit = {
-    val customProperties: ListBuffer[String] = ListBuffer()
-
-    // Collect element custom annotations
-    element.fields.entry(DomainElementModel.CustomDomainProperties) foreach { case FieldEntry(_, v) =>
-      v.value match {
-        case AmfArray(values, _) =>
-          values
-            .sortBy(_.asInstanceOf[DomainExtension].id)
-            .collect {
-              case extension: DomainExtension if !isSemanticExtension(extension) => Some(extension)
-              case _                                                             => None
-            }
-            .flatten
-            .foreach { extension =>
-              val uri = extension.definedBy.id
-              customProperties += uri
-              createCustomExtension(b, uri, extension, None)
-            }
-        case _ => // ignore
-      }
-    }
-
-    // Collect element scalar fields custom annotations
-    var count = 1
-    element.fields.foreach { case (f, v) =>
-      v.value.annotations
-        .collect({ case e: DomainExtensionAnnotation => e })
-        .sortBy(_.extension.id)
-        .foreach(e => {
-          val extension = e.extension
-          val uri       = s"${element.id}/scalar-valued/$count/${extension.name.value()}"
-          customProperties += uri
-          adoptTree(uri, extension.extension) // Fix ids
-          createCustomExtension(b, uri, extension, Some(f))
-          count += 1
-        })
-    }
-
-    if (customProperties.nonEmpty)
-      b.entry(
-          ctx.emitIri(DomainElementModel.CustomDomainProperties.value.iri()),
-          _.list { b =>
-            customProperties.foreach(iri(b, _, inArray = true))
-          }
-      )
-  }
-
-  private def createCustomExtension(
+  override protected def createCustomExtension(
       b: Entry[T],
       uri: String,
       extension: DomainExtension,
@@ -395,9 +299,7 @@ class FlattenedJsonLdEmitter[T](
     )
   }
 
-  private def isSemanticExtension(extension: DomainExtension): Boolean = Option(extension.extension).isEmpty
-
-  def createSortedArray(b: Part[T], seq: Seq[AmfElement], parent: String, element: Type): Unit = {
+  override protected def createSortedArray(b: Part[T], seq: Seq[AmfElement], parent: String, element: Type): Unit = {
     b.obj { b =>
       val id = s"$parent/list"
       createIdNode(b, id)
@@ -409,39 +311,7 @@ class FlattenedJsonLdEmitter[T](
             rb.entry(
                 ctx.emitIri((Namespace.Rdfs + s"_${i + 1}").iri()),
                 { b =>
-                  element match {
-                    case _: Obj =>
-                      e match {
-                        case elementInArray: DomainElement with Linkable if elementInArray.isLink =>
-                          link(b, elementInArray, inArray = true)
-                        case elementInArray: AmfObject =>
-                          obj(b, elementInArray)
-                      }
-                    case Str =>
-                      scalar(b, e, SType.Str)
-
-                    case EncodedIri =>
-                      iri(b, e.asInstanceOf[AmfScalar].toString, inArray = true)
-
-                    case Iri =>
-                      iri(b, e.asInstanceOf[AmfScalar].toString, inArray = true)
-
-                    case Any =>
-                      val scalarElement = e.asInstanceOf[AmfScalar]
-                      scalarElement.value match {
-                        case bool: Boolean =>
-                          typedScalar(b, bool.toString, DataType.Boolean, inArray = true)
-                        case str: String =>
-                          typedScalar(b, str, DataType.String, inArray = true)
-                        case i: Int =>
-                          typedScalar(b, i.toString, DataType.Integer, inArray = true)
-                        case d: Double =>
-                          typedScalar(b, d.toString, DataType.Double, inArray = true)
-                        case f: Float =>
-                          typedScalar(b, f.toString, DataType.Float, inArray = true)
-                        case other => scalar(b, other.toString)
-                      }
-                  }
+                  emitArrayMember(element, e, b)
                 }
             )
           }
@@ -454,146 +324,9 @@ class FlattenedJsonLdEmitter[T](
     }
   }
 
-  private def value(t: Type, v: Value, parent: String, sources: Value => Unit, b: Part[T]): Unit = {
-    t match {
-      case _: ShapeModel if ctx.canGenerateLink(v.value) =>
-        extractToLink(v.value.asInstanceOf[Shape], b)
-      case t: DomainElement with Linkable if t.isLink =>
-        link(b, t)
-        sources(v)
-      case _: Obj =>
-        obj(b, v.value.asInstanceOf[AmfObject])
-        sources(v)
-      case Iri | EncodedIri =>
-        iri(b, v.value.asInstanceOf[AmfScalar].toString)
-        sources(v)
-      case LiteralUri =>
-        typedScalar(b, v.value.asInstanceOf[AmfScalar].toString, DataType.AnyUri)
-        sources(v)
-      case Str =>
-        emitScalar(b, v.value)
-        sources(v)
-      case Bool =>
-        emitScalar(b, v.value, SType.Bool)
-        sources(v)
-      case Type.Int | Type.Long =>
-        emitScalar(b, v.value, SType.Int)
-        sources(v)
-      case Type.Double | Type.Float =>
-        // Doubles get emitted as floats without the @type entry
-        emitScalar(b, v.value, SType.Float)
-        sources(v)
-      case Type.DateTime =>
-        val dateTime = v.value.asInstanceOf[AmfScalar].value.asInstanceOf[SimpleDateTime]
-        typedScalar(b, emitDateFormat(dateTime), DataType.DateTime)
-        sources(v)
-      case Type.Date =>
-        emitDate(v, b)
-        sources(v)
-      case a: SortedArray =>
-        createSortedArray(b, v.value.asInstanceOf[AmfArray].values, parent, a.element)
-        sources(v)
-      case a: Array =>
-        b.list { b =>
-          val seq = v.value.asInstanceOf[AmfArray]
-          sources(v)
-          a.element match {
-            case _: Obj =>
-              seq.values.asInstanceOf[Seq[AmfObject]].foreach {
-                case v @ (_: Shape) if ctx.canGenerateLink(v) =>
-                  extractToLink(v.asInstanceOf[Shape], b, inArray = true)
-                case elementInArray: DomainElement with Linkable if elementInArray.isLink =>
-                  link(b, elementInArray, inArray = true)
-                case elementInArray =>
-                  obj(b, elementInArray)
-              }
-            case Str =>
-              seq.values.asInstanceOf[Seq[AmfScalar]].foreach { e =>
-                scalar(b, e.toString)
-              }
-            case EncodedIri | Iri =>
-              seq.values.asInstanceOf[Seq[AmfScalar]].foreach(e => iri(b, e.toString, inArray = true))
-            case LiteralUri =>
-              typedScalar(b, v.value.asInstanceOf[AmfScalar].toString, DataType.AnyUri, inArray = true)
-            case Type.Int | Type.Long =>
-              seq.values
-                .asInstanceOf[Seq[AmfScalar]]
-                .foreach(e => scalar(b, e.value.toString, SType.Int))
-            case Type.Float =>
-              seq.values
-                .asInstanceOf[Seq[AmfScalar]]
-                .foreach(e => scalar(b, e.value.toString, SType.Float))
-            case Bool =>
-              seq.values
-                .asInstanceOf[Seq[AmfScalar]]
-                .foreach(e => scalar(b, e.value.toString, SType.Bool))
-            case Type.DateTime =>
-              seq.values
-                .asInstanceOf[Seq[AmfScalar]]
-                .foreach { e =>
-                  val dateTime = e.value.asInstanceOf[SimpleDateTime]
-                  typedScalar(b, emitDateFormat(dateTime), DataType.DateTime)
-                }
-            case Type.Date =>
-              seq.values
-                .asInstanceOf[Seq[AmfScalar]]
-                .foreach { e =>
-                  val dateTime = e.value.asInstanceOf[SimpleDateTime]
-                  emitSimpleDateTime(b, dateTime, inArray = false)
-                }
-            case Any =>
-              seq.values.asInstanceOf[Seq[AmfScalar]].foreach { scalarElement =>
-                scalarElement.value match {
-                  case bool: Boolean =>
-                    typedScalar(b, bool.toString, DataType.Boolean, inArray = true)
-                  case i: Int              => typedScalar(b, i.toString, DataType.Integer, inArray = true)
-                  case f: Float            => typedScalar(b, f.toString, DataType.Float, inArray = true)
-                  case d: Double           => typedScalar(b, d.toString, DataType.Double, inArray = true)
-                  case sdt: SimpleDateTime => emitSimpleDateTime(b, sdt)
-                  case other               => scalar(b, other.toString)
-                }
-              }
-            case _ => seq.values.asInstanceOf[Seq[AmfScalar]].foreach(e => iri(b, e.toString, inArray = true))
-          }
-        }
+  override protected def emitArrayObjectMember(b: Part[T], member: AmfObject): Unit = obj(b, member)
 
-      case Any if v.value.isInstanceOf[AmfScalar] =>
-        v.value.asInstanceOf[AmfScalar].value match {
-          case bool: Boolean       => typedScalar(b, bool.toString, DataType.Boolean, inArray = true)
-          case i: Int              => typedScalar(b, i.toString, DataType.Integer, inArray = true)
-          case f: Float            => typedScalar(b, f.toString, DataType.Float, inArray = true)
-          case d: Double           => typedScalar(b, d.toString, DataType.Double, inArray = true)
-          case sdt: SimpleDateTime => emitSimpleDateTime(b, sdt)
-          case other               => scalar(b, other.toString)
-        }
-    }
-  }
-
-  private def emitDate(v: Value, b: Part[T]): Unit = {
-    val maybeDateTime = v.value.asInstanceOf[AmfScalar].value match {
-      case dt: SimpleDateTime => Some(dt)
-      case other              => SimpleDateTime.parse(other.toString).toOption
-    }
-    maybeDateTime match {
-      case Some(dateTime) =>
-        if (dateTime.timeOfDay.isDefined || dateTime.zoneOffset.isDefined) {
-          typedScalar(b, emitDateFormat(dateTime), DataType.DateTime)
-        } else {
-          typedScalar(b, f"${dateTime.year}%04d-${dateTime.month}%02d-${dateTime.day}%02d", DataType.Date)
-        }
-      case _ =>
-        emitScalar(b, v.value)
-    }
-  }
-  private def emitSimpleDateTime(b: Part[T], dateTime: SimpleDateTime, inArray: Boolean = true): Unit = {
-    if (dateTime.timeOfDay.isDefined || dateTime.zoneOffset.isDefined) {
-      typedScalar(b, emitDateFormat(dateTime), DataType.DateTime, inArray)
-    } else {
-      typedScalar(b, f"${dateTime.year}%04d-${dateTime.month}%02d-${dateTime.day}%02d", DataType.Date)
-    }
-  }
-
-  private def obj(b: Part[T], obj: AmfObject): Unit = {
+  override protected def obj(b: Part[T], obj: AmfObject, inArray: Boolean = false): Unit = {
     def emit(b: Part[T]): Unit = {
       b.obj(createIdNode(_, obj.id))
       pending.tryEnqueue(obj)
@@ -601,199 +334,47 @@ class FlattenedJsonLdEmitter[T](
     emit(b)
   }
 
-  private def extractToLink(shape: Shape, b: Part[T], inArray: Boolean = false): Unit = {
-    if (!ctx.isDeclared(shape) && !ctx.isInReferencedShapes(shape)) {
-      ctx + shape
-      shape.name.option() match {
-        case None =>
-          shape.withName("inline-type")
-          shape.annotations += InlineElement()
-        case Some("schema") | Some("type") =>
-          shape.annotations += InlineElement()
-        case _ if !shape.annotations.contains(classOf[DeclaredElement]) =>
-          shape.annotations += InlineElement() // to catch media type named cases.
-        case _ => // ignore
+  override protected def emitAmfObject(amfObj: AmfObject, entry: Entry[T]): Unit = {
+    createIdNode(entry, amfObj.id)
+    pending.tryEnqueue(amfObj)
+  }
+
+
+
+  override protected def emitAnnotations(id: String, filteredSources: SourceMap, b: Entry[T]): Unit = {
+    createIdNode(b, id)
+    val e = new Emission((part: Part[T]) => {
+      part.obj { rb =>
+        createIdNode(rb, id)
+        createTypeNode(rb, SourceMapModel)
+        createAnnotationNodes(id, rb, filteredSources.annotations)
+        createAnnotationNodes(id, rb, filteredSources.eternals)
       }
-    }
-    val linkLabel = shape.name.value()
-    val linked = shape match {
-      // if it is recursive we force the conversion into a linked shape
-      case rec: RecursiveShape =>
-        val hash = s"${rec.id}$linkLabel".hashCode
-        RecursiveShape()
-          .withId(s"${rec.id}/link-$hash")
-          .withLinkTarget(rec)
-          .withLinkLabel(linkLabel)
-      // no recursive we just generate the linked shape
-      case _ =>
-        shape.link[Shape](linkLabel)
-    }
-
-    link(b, linked, inArray)
+    }) with Metadata
+    e.id = Some(id)
+    e.isDeclaration = ctx.emittingDeclarations
+    e.isReference = ctx.emittingReferences
+    pending.tryEnqueue(e)
   }
 
-  private def link(b: Part[T], elementWithLink: DomainElement with Linkable, inArray: Boolean = false): Unit = {
-    def emit(b: Part[T]): Unit = {
-      // before emitting, we remove the link target to avoid loops and set
-      // the fresh value for link-id
-      val savedLinkTarget = elementWithLink.linkTarget
-      elementWithLink.linkTarget.foreach { target =>
-        elementWithLink.set(LinkableElementModel.TargetId, target.id)
-        elementWithLink.fields.removeField(LinkableElementModel.Target)
+  override protected def emitEternalsNode(id: String, sources: SourceMap, b: Entry[T]): Unit = {
+    createIdNode(b, id)
+    val e = new Emission((part: Part[T]) => {
+      part.obj { rb =>
+        createIdNode(rb, id)
+        createTypeNode(rb, SourceMapModel)
+        createAnnotationNodes(id, rb, sources.eternals)
       }
+    }) with Metadata
 
-      b.obj { o =>
-        createIdNode(o, elementWithLink.id)
-        pending.tryEnqueue(elementWithLink)
-      }
-      // we reset the link target after emitting
-      savedLinkTarget.foreach { target =>
-        elementWithLink.fields.setWithoutId(LinkableElementModel.Target, target)
-      }
-    }
-
-    if (inArray) emit(b) else b.list(emit)
+    e.id = Some(id)
+    e.isDeclaration = ctx.emittingDeclarations
+    e.isReference = ctx.emittingReferences
+    pending.tryEnqueue(e)
   }
 
-  private def iri(b: Part[T], content: String, inArray: Boolean = false): Unit = {
-    def emit(b: Part[T]): Unit = {
-      b.obj(_.entry(JsonLdKeywords.Id, raw(_, ctx.emitId(content))))
-    }
-
-    if (inArray) emit(b) else b.list(emit)
-  }
-
-  private def typedScalar(b: Part[T], content: String, dataType: String, inArray: Boolean = false): Unit = {
-    def emit(b: Part[T]): Unit = b.obj { m =>
-      m.entry(JsonLdKeywords.Value, raw(_, content))
-      m.entry(JsonLdKeywords.Type, raw(_, ctx.emitIri(dataType)))
-    }
-
-    if (inArray) emit(b) else b.list(emit)
-  }
-
-  private def createIdNode(b: Entry[T], id: String): Unit = b.entry(
-      JsonLdKeywords.Id,
-      raw(_, ctx.emitId(id))
-  )
-
-  private def createSourcesNode(id: String, sources: SourceMap, b: Entry[T]): Unit = {
-    val filteredSources  = if(options.governanceMode) filterSourceMaps(sources) else sources
-    val withSourceMaps   = options.isWithSourceMaps || options.governanceMode
-
-    if (withSourceMaps && filteredSources.nonEmpty) {
-      if (options.isWithRawSourceMaps) {
-        b.entry(
-            "smaps",
-            _.obj { b =>
-              createAnnotationNodes(id, b, filteredSources.annotations)
-              createAnnotationNodes(id, b, filteredSources.eternals)
-            }
-        )
-      } else {
-        b.entry(
-            ctx.emitIri(DomainElementModel.Sources.value.iri()),
-            _.list {
-              _.obj { b =>
-                createIdNode(b, id)
-                val e = new Emission((part: Part[T]) => {
-                  part.obj { rb =>
-                    createIdNode(rb, id)
-                    createTypeNode(rb, SourceMapModel)
-                    createAnnotationNodes(id, rb, filteredSources.annotations)
-                    createAnnotationNodes(id, rb, filteredSources.eternals)
-                  }
-                }) with Metadata
-                e.id = Some(id)
-                e.isDeclaration = ctx.emittingDeclarations
-                e.isReference = ctx.emittingReferences
-                pending.tryEnqueue(e)
-              }
-            }
-        )
-      }
-    } else {
-      createEternalsAnnotationsNodes(id, options, b, filteredSources)
-    }
-  }
-
-  private def filterSourceMaps(sources: SourceMap) = {
-    val filteredAnnotations = sources.annotations.filter(a => allowList.contains(a._1))
-    val filteredEternals    = sources.eternals.filter(a => allowList.contains(a._1))
-    new SourceMap(filteredAnnotations, filteredEternals)
-  }
-
-  private def createEternalsAnnotationsNodes(
-      id: String,
-      options: RenderOptions,
-      b: Entry[T],
-      sources: SourceMap
-  ): Unit = {
-    if (sources.eternals.nonEmpty)
-      if (options.isWithRawSourceMaps) {
-        b.entry(
-            "smaps",
-            _.obj { b =>
-              createAnnotationNodes(id, b, sources.eternals)
-            }
-        )
-      } else {
-        b.entry(
-            ctx.emitIri(DomainElementModel.Sources.value.iri()),
-            _.list {
-              _.obj { b =>
-                createIdNode(b, id)
-                val e = new Emission((part: Part[T]) => {
-                  part.obj { rb =>
-                    createIdNode(rb, id)
-                    createTypeNode(rb, SourceMapModel)
-                    createAnnotationNodes(id, rb, sources.eternals)
-                  }
-                }) with Metadata
-
-                e.id = Some(id)
-                e.isDeclaration = ctx.emittingDeclarations
-                e.isReference = ctx.emittingReferences
-                pending.tryEnqueue(e)
-              }
-            }
-        )
-      }
-  }
-
-  private def createAnnotationNodes(
-      id: String,
-      b: Entry[T],
-      annotations: mutable.ListMap[String, mutable.ListMap[String, String]]
-  ): Unit = {
-    annotations.foreach({ case (a, values) =>
-      if (ctx.options.isWithRawSourceMaps) {
-        b.entry(
-            a,
-            _.obj { o =>
-              values.foreach { case (iri, v) =>
-                o.entry(
-                    ctx.emitId(ctx.emitIri(iri)),
-                    raw(_, v)
-                )
-              }
-            }
-        )
-      } else {
-        b.entry(
-            ctx.emitIri(ValueType(Namespace.SourceMaps, a).iri()),
-            _.list(b =>
-              values.zipWithIndex.foreach { case (tuple, index) =>
-                createAnnotationValueNode(s"$id/$a/element_$index", b, tuple)
-              }
-            )
-        )
-      }
-    })
-  }
-
-  private def createAnnotationValueNode(id: String, b: Part[T], tuple: (String, String)): Unit =
-    tuple match {
+  override protected def createAnnotationValueNode(id: String, b: Part[T], annotationEntry: (String, String)): Unit =
+    annotationEntry match {
       case (iri, v) =>
         b.obj { b =>
           createIdNode(b, id)
@@ -811,19 +392,10 @@ class FlattenedJsonLdEmitter[T](
         pending.tryEnqueue(e)
     }
 
-  private def emitDateFormat(dateTime: SimpleDateTime) = dateTime.toString
+  override protected def scalar(b: Part[T], content: String, t: SType): Unit = b += Scalar(t, content)
 
-  private def scalar(b: Part[T], content: String, t: SType): Unit = b += Scalar(t, content)
+  override protected def emitScalar(b: Part[T], content: String, t: SType): Unit = scalar(b, content, t)
 
-  private def scalar(b: Part[T], content: String): Unit = scalar(b, content, SType.Str)
-
-  private def scalar(b: Part[T], content: AmfElement, t: SType): Unit =
-    scalar(b, content.asInstanceOf[AmfScalar].value.toString, t)
-
-  private def emitScalar(b: Part[T], content: String, t: SType): Unit = scalar(b, content, t)
-  private def emitScalar(b: Part[T], content: String): Unit           = emitScalar(b, content, SType.Str)
-
-  private def emitScalar(b: Part[T], content: AmfElement, t: SType): Unit = scalar(b, content, t)
-  private def emitScalar(b: Part[T], content: AmfElement): Unit           = emitScalar(b, content, SType.Str)
+  override protected def emitScalar(b: Part[T], content: AmfElement, t: SType): Unit = scalar(b, content, t)
 
 }
