+++ shared/src/main/scala/amf/core/internal/plugins/document/graph/emitter/InstanceMetaApplicableFieldRenderProvider.scala
@@ -1,0 +1,10 @@
+package amf.core.internal.plugins.document.graph.emitter
+
+import amf.core.client.scala.config.RenderOptions
+import amf.core.client.scala.model.domain.AmfObject
+import amf.core.internal.metamodel.Field
+
+object InstanceMetaApplicableFieldRenderProvider extends ApplicableMetaFieldRenderProvider {
+  override def fieldsFor(element: AmfObject, renderOptions: RenderOptions): Seq[Field] =
+    super.fieldsFor(element, renderOptions)
+}
