+++ shared/src/main/scala/amf/core/internal/plugins/document/graph/emitter/utils/SourceMapEmitter.scala
@@ -1,0 +1,13 @@
+package amf.core.internal.plugins.document.graph.emitter.utils
+
+trait SourceMapEmitter {
+  protected def sourceMapIdFor(id: String): String = {
+    if (id.endsWith("/")) {
+      id + "source-map"
+    } else if (id.contains("#") || id.startsWith("null")) {
+      id + "/source-map"
+    } else {
+      id + "#/source-map"
+    }
+  }
+}
